Fix bottleneck issue of rand() function while generating random print character
While running the ycsbc benchmark code, we have observed that rand() function become the bottleneck issue for dram code while running in parallel threads. It's a general case of rand() function for parallel thread case due to the thread safety nature of rand(). The surprising thing is it don't perform bad in pmem code. Will investigate the issue later. For now, I have replaced the rand() function by rand_r() and generate the seed in workload init cycle. So for different init call, the seed will be different.