Fix issue of seed generator for rand_r function
If rand_r() is called with the same initial value for the object pointed to by seed and that object is not modified between successive returns and calls to rand_r(), the same sequence shall be generated. So while calling rand_r() function, updated the seed value. This again will hamper the benchmarking performance. We will need to find an alternative way for this.