Fix issue regarding finding key in b-plus tree
Internal nodes of b-plus tree will contain the key that also exist in the leaf as well. So, if the expected key is found in internal nodes, we need to go for the next child. This is one of the differentiating point between b-tree and b-plus tree. Also added the value along with the key while printing the leaf chain.